api-gateway/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
httpx
python-dotenv
